fix(lib-client): resolve bool/narrowing and signed-unsigned warnings in game stage, skill basic, taxi station, and dxincommand

+++ Lib_ClientUI/Interface/UIItemInfoLoader.cpp
@@ -1036,7 +1036,7 @@ namespace NS_ITEMINFO
 					strText.Format("%s:%s", ID2GAMEWORD("ITEM_BASIC_INFO", 1), strMoney);
 
 					//	커미션 액수
-					dwCOMMISSION_MONEY = dwPrice - dwNpcSellPrice;
+					dwCOMMISSION_MONEY = static_cast<DWORD>(dwPrice - dwNpcSellPrice);
 
 					D3DCOLOR dwColor = NS_UITEXTCOLOR::RED;
 					if (dwPrice <= GLGaeaClient::GetInstance().GetCharacterLogic().m_lnMoney)
